+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-01-24
+
+### Added
+
+- Added `getField` method to `AbstractForm`
+- Added `phpstan/phpstan` to help improve code quality
+
+### Changed
+
+- N/A
+
+### Deprecated
+
+- N/A
+
+### Removed
+
+- N/A
+
+### Fixed
+
+- N/A
+
+### Security
+
+- N/A
+
 ## [0.2.0] - 2024-11-11
 
 ### Added
@@ -12,14 +39,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Added custom form field validation using a callback
 - Added `SubmitButton` field type
 
-### Fixed
-
-- N/A
-
 ### Changed
 
 - N/A
 
+### Deprecated
+
+- N/A
+
 ### Removed
+
+- N/A
+
+### Fixed
+
+- N/A
+
+### Security
 
 - N/A